% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccwarcs-package.R
\docType{package}
\name{ccwarcs-package}
\alias{ccwarcs}
\alias{ccwarcs-package}
\title{ccwarcs: Access to Common Crawl WARC Files via AWS}
\description{
Provides access to Common Crawl WARC files via
    Amazon Web Services.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jasonmtroos/ccwarcs}
  \item Report bugs at \url{https://github.com/jasonmtroos/ccwarcs/issues}
}

}
\author{
\strong{Maintainer}: Jason Roos \email{my_first_initial_and_last_name@rsm.nl}

}
\keyword{internal}
