% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ccwarcs_options}
\alias{ccwarcs_options}
\title{Set Options for the ccwarcs Package}
\usage{
ccwarcs_options(cache = "~/.ccwarcs_cache", cdx_sleep = 0.3,
  page_size = NULL)
}
\arguments{
\item{cache}{Location for the cached CDX index results and WARC files}

\item{cdx_sleep}{Amount of time (seconds) to wait after a CDX server call before resuming}

\item{page_size}{An optional parameter to set the amount of data returned in each page}
}
\value{
A list object of class \code{ccwarcs_options}
}
\description{
Set Options for the ccwarcs Package
}
\examples{
\dontrun{
opts <- ccwarcs_options(cache = "my_project_directory")
url <- "http://www.celebuzz.com/2017-01-04"
crawl <- "2018-47"
results <- get_cc_index(url, crawl, .options = opts)
}
}
