% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdx.R
\name{cdx_fetch_list_of_indexes}
\alias{cdx_fetch_list_of_indexes}
\title{Get a List of Indexes from the Common Crawl Index Server}
\usage{
cdx_fetch_list_of_indexes()
}
\value{
A tibble
}
\description{
Get a List of Indexes from the Common Crawl Index Server
}
\examples{
# not run:
# cdx_fetch_list_of_indexes()
}
