% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdx.R
\name{get_cc_index}
\alias{get_cc_index}
\title{Get a List of WARCs from Common Crawl Index Server}
\usage{
get_cc_index(urls, crawls, .options = NULL)
}
\arguments{
\item{urls}{A vector of URLs of captured pages, allowing \code{*} as a wildcard character}

\item{crawls}{A vector of Ids of CC crawls to search

Values in \code{crawls} are typically character strings in the format \code{YYYY-ww}, e.g.
2018-47 for the crawl published in the 47th week of 2018.
See \url{https://index.commoncrawl.org/} for a list of crawls,
and \link{cdx_fetch_list_of_crawls} for programmatic access
to this list.}

\item{.options}{An optional object of class \link{ccwarcs_options}}
}
\value{
A tibble
}
\description{
The list of WARCs is cached in a directory specified by the \code{cache}  argument to \link{ccwarcs_options}
}
\examples{
# not run:
# url <- "http://www.celebuzz.com/2017-01-04"
# crawl <- "2018-47"
# results <- get_cc_index(url, crawl)
}
