% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warcs.R
\name{test_AWS_credentials}
\alias{test_AWS_credentials}
\title{Test AWS credentials}
\usage{
test_AWS_credentials()
}
\description{
A helper function that wraps \link[aws.signature:locate_credentials]{locate_credentials} and provides
advice if AWS credentials cannot be found.
}
\seealso{
\link[aws.signature:locate_credentials]{locate_credentials}
}
