% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warcs.R
\name{get_warc}
\alias{get_warc}
\title{Get a WARC from the Common Crawl via AWS}
\usage{
get_warc(filename, offset, length, digest, include_headers = FALSE)
}
\arguments{
\item{filename}{AWS path to the WARC}

\item{offset}{Starting byte offset for the chunk}

\item{length}{Number of bytes in the chunk}

\item{digest}{Common Crawl digest for requested chunk}

\item{include_headers}{If TRUE, include the WARC and HTTP headers in the result. See below.}
}
\value{
HTML contents of the requested WARC, and optionally the WARC and HTTP headers.
If \code{include_headers = TRUE}, the result is a list with elements \code{headers} and \code{payload}.
Otherwise the result is a character string (vector) containing the HTML of the requested WARC.
}
\description{
The WARC is cached in a directory specified by the \code{cache}  argument to \link{ccwarcs_options}
}
